Added Terms and condition and update ui

+++ wp-content/plugins/simply-static/src/tasks/class-ss-create-zip-archive.php
@@ -47,6 +47,12 @@ class Create_Zip_Archive_Task extends Task {
 		if ( ! $temp_dir_empty ) {
 			foreach ( new \DirectoryIterator( $temp_dir ) as $file ) {
 				if ( ! $file->isDir() ) {
+					$can_delete_file = apply_filters( 'ss_can_delete_file', true, $file, $temp_dir );
+
+					if ( ! $can_delete_file ) {
+						continue;
+					}
+
 					unlink( $file->getPathname() );
 				}
 			}
